Update docs and tests based on feedback
